Add files via upload
The Cud_inDrive_DB is an application developed using the Tkinter framework in Python. It allows users on the same Wi-Fi network to share and collect photos, which are then stored in a MySQL database on the host PC. The application offers a user-friendly interface for effortless photo collection and database management.